crypto + io + circuits + execution: SSWU BN_mod_* checks, len/length guards, retry cap, tighter Backend encapsulation
crypto/ec.cpp map_to_curve_sswu_p256 now checks every BN_mod_sqr /
BN_mod_mul / BN_mod_add / BN_sub / BN_copy return value. BN_mod_*
failures leave the destination BIGNUM in an indeterminate state — not
null — so downstream code previously read garbage instead of
crashing. Comment above the function updated to match.

io/net_io_channel.h + io/tls_io_channel.h: send_data_internal /
recv_data_internal reject negative len at entry. A negative int64_t
flowing into memcpy(..., len) would coerce to a huge size_t and copy
gigabytes / UB.

crypto/prg.h random_data_unaligned and crypto/ccrh.h CCRH::Hn
early-return on nbytes/length <= 0. The PRG path otherwise hit
memcpy(..., nbytes) with size_t coercion; CCRH::Hn otherwise
allocated ~SIZE_MAX via new block[length].

circuits/circuit_file.h: the int-arg BristolFormat ctor now casts
num_gate to size_t before multiplying by 4, avoiding signed overflow
on large circuits (num_gate > INT_MAX/4). BristolFormat::compute and
BristolFashion::compute now document the caller's array-size
precondition since neither bounds-checks the file-declared fan-in/out
against the caller's allocation.

io/tcp_socket.h: server_listen and client_connect now check the
::socket() return value, and client_connect caps retries at ~60 s
total (60000 × 1 ms) before erroring out rather than hanging
indefinitely on a permanently-unreachable peer.

execution/half_gate.h + execution/privacy_free.h: MITCCRH<8> mitccrh
and PRP prp moved from public to protected. The mix-mode invariant
documented in mitccrh.h (tweak-mode and gid-mode must not mix on one
instance) is enforceable only if external code can't reach the
field; subclasses HalfGateGen / HalfGateEva / PrivacyFreeGen /
PrivacyFreeEva still have access.

Also standardize makeBlock(2*idx, 0ULL) in privacy_free.h — the prior
(long)0 vs (uint64_t)0 inconsistency was platform-sensitive on LP64
vs LLP64 hosts.

+++ emp-tool/circuits/circuit_file.h
@@ -63,8 +63,8 @@ class BristolFormat { public:
 		this->n1 = n1;
 		this->n2 = n2;
 		this->n3 = n3;
-		gates.resize(num_gate*4);
-		memcpy(gates.data(), gate_arr, num_gate*4*sizeof(int));
+		gates.resize((size_t)num_gate * 4);
+		memcpy(gates.data(), gate_arr, (size_t)num_gate * 4 * sizeof(int));
 	}
 
 	BristolFormat(FILE * file) {
@@ -146,6 +146,11 @@ class BristolFormat { public:
 		fclose(f);
 	}
 
+	// Precondition: caller's `in1`, `in2`, `out` must hold at least `n1`,
+	// `n2`, `n3` elements respectively. No bounds check is performed —
+	// a malformed circuit file with mismatched fan-in/out vs. caller
+	// allocation reads/writes OOB. n1/n2/n3 are the file-declared
+	// bit-widths exposed as public fields, so callers can size accordingly.
 	template<typename Wire>
 	void compute(Bit_T<Wire> * out, const Bit_T<Wire>* in1, const Bit_T<Wire> * in2) {
 		vector<Bit_T<Wire>> wires(num_wire);
@@ -239,6 +244,8 @@ class BristolFashion { public:
 		fclose(f);
 	}
 
+	// Precondition: caller's `in` and `out` must hold at least `num_input`
+	// and `num_output` elements respectively. No bounds check is performed.
 	template<typename Wire>
 	void compute(Bit_T<Wire> * out, const Bit_T<Wire> * in) {
 		vector<Bit_T<Wire>> wires(num_wire);
